pyproject.toml
[project]
name = "mlex_dimension_reduction_pca"
version = "0.0.2"
authors = [
    { name="Runbo Jiang", email="rjiang2@lbl.gov"}
]
description = "PCA algorithm for dimension reduction."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "dask>=2024.3.0",
    "dask-expr==1.0.10",
    "pydantic",
    "numpy",
    "Pillow",
    "joblib",
    "scikit-learn",
    "pandas",
    "pyarrow",
    "tiled[client]==0.1.0a118",
]

[project.optional-dependencies]
# These are required for developing the package (running the tests, building
# the documentation) but not necessarily required for _using_ it.
dev = [
    "flake8",
    "pre-commit",
    "pytest"
]

[project.urls]
Homepage = "https://github.com/mlexchange/mlex_dimension_reduction_pca"
Issues = "https://github.com/mlexchange/mlex_dimension_reduction_pca/issues/"

[tool.isort]
profile = "black"
